Closed Testing Ready

+++ Assets/Scripts/Data Scripts/GameData.cs
@@ -18,7 +18,7 @@ public class ObjectivesProgress
     public readonly List<int> objectiveRewards = new List<int> { 1, 1, 1, 3, 3, 3, 5, 5, 5, 15 };
     public readonly List<int> scoreGoals = new List<int> { 10, 25, 50, 100, 250, 500, 750, 1000, 2500, 5000 };
     public readonly List<int> coinGoals = new List<int> { 5, 15, 25, 50, 75, 100, 150, 250, 500, 1000, 1500 };
-    public readonly List<float> timeGoals = new List<float> { 60.0f, 300.0f, 900.0f, 1800.0f, 3600.0f, 21600.0f, 43200.0f, 86400.0f, 604800.0f, 2592000.0f };
+    public readonly List<int> timeGoals = new List<int> { 60, 300, 900, 1800, 3600, 21600, 43200, 86400, 604800, 2592000 };
     public int scoreObjectiveStep = 0;
     public int coinObjectiveStep = 0;
     public int timeObjectiveStep = 0;
